Relocation to webpack
