+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules/
-build/
+docs/
 npm-debug.*
