+++ package.json
@@ -6,10 +6,14 @@
   "scripts": {
     "editorconfig": "editorconfig",
     "stylelint": "stylelint \"source/sass/**/*.scss\" --syntax scss",
-    "test": "npm run editorconfig && npm run stylelint",
-    "build": "gulp build",
-    "start": "gulp start"
+    "check": "npm run editorconfig && npm run stylelint",
+    "build": "webpack --mode production",
+    "dev": "webpack-dev-server --mode development"
   },
+  "browserslist": [
+    "> 1%",
+    "last 3 version"
+  ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/PapkaRimskyi/watch-shop.git"
@@ -21,26 +25,28 @@
   },
   "homepage": "https://github.com/PapkaRimskyi/watch-shop#readme",
   "devDependencies": {
+    "@babel/core": "^7.6.4",
     "@babel/preset-env": "^7.5.5",
     "autoprefixer": "^9.5.1",
-    "babel-core": "^6.26.3",
-    "browser-sync": "^2.26.5",
-    "del": "^4.1.1",
+    "babel-loader": "^8.0.6",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^5.0.4",
+    "css-loader": "^3.2.0",
+    "cssnano": "^4.1.10",
     "editorconfig": "^0.15.3",
-    "gulp": "^4.0.1",
-    "gulp-babel": "^8.0.0",
-    "gulp-concat": "^2.6.1",
-    "gulp-csso": "^3.0.1",
-    "gulp-imagemin": "^5.0.3",
-    "gulp-plumber": "^1.2.1",
-    "gulp-postcss": "^8.0.0",
-    "gulp-rename": "^1.4.0",
-    "gulp-sass": "^4.0.2",
-    "gulp-uglify": "^3.0.2",
-    "gulp-webp": "^4.0.1",
+    "file-loader": "^4.2.0",
+    "html-webpack-plugin": "^3.2.0",
+    "image-webpack-loader": "^6.0.0",
+    "mini-css-extract-plugin": "^0.8.0",
+    "node-sass": "^4.12.0",
+    "postcss-loader": "^3.0.0",
+    "sass-loader": "^8.0.0",
+    "style-loader": "^1.0.0",
     "stylelint": "^10.0.1",
     "stylelint-config-recommended": "^2.2.0",
-    "svgo": "^1.2.2"
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.9",
+    "webpack-dev-server": "^3.9.0"
   },
   "editorconfig": [
     "*.json",
