+++ postcss.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  plugins: [
+    require('autoprefixer')({
+      grid: 'autoplace',
+    }),
+    require('cssnano')({
+      preset: 'default',
+    }),
+  ]
+}
