+++ source/catalog.html
@@ -6,14 +6,14 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <meta name="description" content="Интернет-магазин часов Conquest">
   <title>Каталог часов Conquest</title>
-  <link rel="stylesheet" href="css/style.css">
-  <link rel="preload" href="fonts/PlayfairDisplay-Regular.woff" as="font" crossorigin="anonymous">
-  <link rel="preload" href="fonts/PlayfairDisplay-Bold.woff" as="font" crossorigin="anonymous">
-  <link rel="preload" href="fonts/Prata-Regular.woff" as="font" crossorigin="anonymous">
-  <link rel="preload" href="fonts/Roboto-Light.woff" as="font" crossorigin="anonymous">
-  <link rel="preload" href="fonts/Roboto-Medium.woff" as="font" crossorigin="anonymous">
-  <link rel="preload" href="fonts/Roboto-Bold.woff" as="font" crossorigin="anonymous">
-  <link rel="shortcut icon" href="img/favicon.ico" type="image/x-icon">
+  <link rel="stylesheet" href="assets/css/style.css">
+  <link rel="preload" href="assets/fonts/PlayfairDisplay-Regular.woff" as="font" crossorigin="anonymous">
+  <link rel="preload" href="assets/fonts/PlayfairDisplay-Bold.woff" as="font" crossorigin="anonymous">
+  <link rel="preload" href="assets/fonts/Prata-Regular.woff" as="font" crossorigin="anonymous">
+  <link rel="preload" href="assets/fonts/Roboto-Light.woff" as="font" crossorigin="anonymous">
+  <link rel="preload" href="assets/fonts/Roboto-Medium.woff" as="font" crossorigin="anonymous">
+  <link rel="preload" href="assets/fonts/Roboto-Bold.woff" as="font" crossorigin="anonymous">
+  <link rel="shortcut icon" href="assets/img/favicon.ico" type="image/x-icon">
 </head>
 <body>
   <svg style="display: none;">
@@ -403,7 +403,6 @@
       </div>
     </li>
   </template>
-  <script src="js/mainJs.js"></script>
-  <script src="js/catalog.js"></script>
+  <script src="assets/js/catalog.js"></script>
 </body>
 </html>
