+++ source/index.html
@@ -6,14 +6,14 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <meta name="description" content="Интернет-магазин часов Conquest">
   <title>Conquest - интернет-магазин часов</title>
-  <link rel="stylesheet" href="css/style.css">
-  <link rel="preload" href="fonts/PlayfairDisplay-Regular.woff" as="font" crossorigin="anonymous">
-  <link rel="preload" href="fonts/PlayfairDisplay-Bold.woff" as="font" crossorigin="anonymous">
-  <link rel="preload" href="fonts/Prata-Regular.woff" as="font" crossorigin="anonymous">
-  <link rel="preload" href="fonts/Roboto-Light.woff" as="font" crossorigin="anonymous">
-  <link rel="preload" href="fonts/Roboto-Medium.woff" as="font" crossorigin="anonymous">
-  <link rel="preload" href="fonts/Roboto-Bold.woff" as="font" crossorigin="anonymous">
-  <link rel="shortcut icon" href="img/favicon.ico" type="image/x-icon">
+  <link rel="stylesheet" href="assets/css/style.css">
+  <link rel="preload" href="assets/fonts/PlayfairDisplay-Regular.woff" as="font" crossorigin="anonymous">
+  <link rel="preload" href="assets/fonts/PlayfairDisplay-Bold.woff" as="font" crossorigin="anonymous">
+  <link rel="preload" href="assets/fonts/Prata-Regular.woff" as="font" crossorigin="anonymous">
+  <link rel="preload" href="assets/fonts/Roboto-Light.woff" as="font" crossorigin="anonymous">
+  <link rel="preload" href="assets/fonts/Roboto-Medium.woff" as="font" crossorigin="anonymous">
+  <link rel="preload" href="assets/fonts/Roboto-Bold.woff" as="font" crossorigin="anonymous">
+  <link rel="shortcut icon" href="assets/img/favicon.ico" type="image/x-icon">
 </head>
 <body>
   <svg style="display: none;">
@@ -153,7 +153,7 @@
       <div class="new-collection-preview__product">
         <p class="new-collection-preview__product-price">22 000 Р</p>
         <div class="new-collection-preview__product-img-container">
-          <img class="new-collection-preview__product-img" src="img/rado-watch.png" width="402" height="795" alt="Часы Rado">
+          <img class="new-collection-preview__product-img" src="assets/img/rado-watch.png" width="402" height="795" alt="Часы Rado">
         </div>
         <p class="new-collection-preview__product-diameter">28 мм диаметр</p>
       </div>
@@ -170,7 +170,7 @@
               <span class="popular-models__product-info popular-models__product-info--model">Rado</span><br>
               65 300 Р
             </p>
-            <img class="popular-models__product-img" src="img/watch_1.png" alt="Фотография часов" width="309" height="318">
+            <img class="popular-models__product-img" src="assets/img/watch_1.png" alt="Фотография часов" width="309" height="318">
             <a href="#" class="popular-models__product-link">Подробнее</a>
           </div>
         </div>
@@ -180,7 +180,7 @@
               <span class="popular-models__product-info popular-models__product-info--model">Rado</span><br>
               65 300 Р
             </p>
-            <img class="popular-models__product-img" src="img/watch_2.png" alt="Фотография часов" width="490" height="492">
+            <img class="popular-models__product-img" src="assets/img/watch_2.png" alt="Фотография часов" width="490" height="492">
             <a href="#" class="popular-models__product-link">Подробнее</a>
           </div>
         </div>
@@ -190,7 +190,7 @@
               <span class="popular-models__product-info popular-models__product-info--model">Bvlgari</span><br>
               81 700 Р
             </p>
-            <img class="popular-models__product-img" src="img/watch_3.png" alt="Фотография часов" width="274" height="240">
+            <img class="popular-models__product-img" src="assets/img/watch_3.png" alt="Фотография часов" width="274" height="240">
             <a href="#" class="popular-models__product-link">Подробнее</a>
           </div>
         </div>
@@ -200,7 +200,7 @@
               <span class="popular-models__product-info popular-models__product-info--model">Bvlgari</span><br>
               72 700 Р
             </p>
-            <img class="popular-models__product-img popular-models__product-img--position-coord" src="img/watch_4.png" alt="Фотография часов" width="245" height="246">
+            <img class="popular-models__product-img popular-models__product-img--position-coord" src="assets/img/watch_4.png" alt="Фотография часов" width="245" height="246">
             <a href="#" class="popular-models__product-link">Подробнее</a>
           </div>
         </div>
@@ -224,7 +224,7 @@
       <div class="advantages__collection-container">
         <div class="advantages__slider-container advantages__slider-container--active">
           <div class="advantages__img-container">
-            <img class="advantages__img" src="img/clockBanner.jpg" alt="Часы" width="521" height="414">
+            <img class="advantages__img" src="assets/img/clockBanner.jpg" alt="Часы" width="521" height="414">
           </div>
           <div class="advantages__description">
             <h3 class="advantages__description-headline">Подлинное качество</h3>
@@ -234,7 +234,7 @@
         </div>
         <div class="advantages__slider-container">
           <div class="advantages__img-container">
-            <img class="advantages__img" src="img/clockBanner.jpg" alt="Часы" width="521" height="414">
+            <img class="advantages__img" src="assets/img/clockBanner.jpg" alt="Часы" width="521" height="414">
           </div>
           <div class="advantages__description">
             <h3 class="advantages__description-headline">Подлинное качество 2</h3>
@@ -288,9 +288,9 @@
           </ul>
         </div>
         <div class="social__clock-illustration">
-          <img class="social__clock-illustration-photo social__clock-illustration-photo--small" src="img/illustration-clock1.jpg" alt="Демонстрация наручных часов" width="242" height="228">
-          <img class="social__clock-illustration-photo social__clock-illustration-photo--big" src="img/illustration-clock3.jpg" alt="Демонстрация наручных часов" width="377" height="460">
-          <img class="social__clock-illustration-photo social__clock-illustration-photo--small" src="img/illustration-clock2.jpg" alt="Демонстрация наручных часов" width="242" height="222">
+          <img class="social__clock-illustration-photo social__clock-illustration-photo--small" src="assets/img/illustration-clock1.jpg" alt="Демонстрация наручных часов" width="242" height="228">
+          <img class="social__clock-illustration-photo social__clock-illustration-photo--big" src="assets/img/illustration-clock3.jpg" alt="Демонстрация наручных часов" width="377" height="460">
+          <img class="social__clock-illustration-photo social__clock-illustration-photo--small" src="assets/img/illustration-clock2.jpg" alt="Демонстрация наручных часов" width="242" height="222">
         </div>
       </div>
     </section>
@@ -375,7 +375,6 @@
       </div>
     </section>
   </footer>
-  <script src="js/mainJs.js"></script>
-  <script src="js/index.js"></script>
+  <script src="assets/js/index.js"></script>
 </body>
 </html>
