+++ source/styles/blocks/advantages.scss
