+++ source/styles/blocks/animation.scss
