+++ source/styles/blocks/catalog-name-and-sort.scss
