+++ source/styles/blocks/filter-and-product.scss
