+++ source/styles/blocks/footer.scss
