+++ source/styles/blocks/information.scss
