+++ source/styles/blocks/new-collection-preview.scss
