+++ source/styles/blocks/product.scss
