+++ source/styles/blocks/slider-button.scss
