+++ source/styles/blocks/user-collection.scss
